requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.35
pymysql==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic-settings==2.4.0
httpx==0.27.2
pytest==8.3.3
python-multipart==0.0.9